+++ kgql-gradle-plugin/src/main/groovy/com/codingfeline/kgql/gradle/KgqlTask2.groovy
@@ -35,7 +35,6 @@ class KgqlTask2 extends SourceTask {
         outputDirectory.mkdirs()
 
         def environment = new KgqlEnvironment(
-                sourceFolders.findAll { it.exists() }.toList(),
                 getSource().toList(),
                 packageName,
                 outputDirectory,